# openapi.yaml
openapi: 3.0.3
info:
  title: ToolBox JL — API Platform
  description: >
    Contrato de API canónico de la plataforma ToolBox JL (gestión, alquiler, venta de
    herramientas y logística inteligente). Este archivo es un artefacto de
    Spec-Driven Development (SDD): se escribe y versiona ANTES que el código
    (spec-first) y es la fuente de verdad que Claude Code consume para generar e
    implementar los controladores, DTOs y guards de rol del backend NestJS. El
    código se valida por conformidad contra este contrato, no al revés.

    Fuente de contenido: Esquema de Backend (DOC-BE-01) §3.2 (diccionario de
    datos) y §4 (contratos de API por módulo); complementado con TRD §4 para las
    rutas invocadas por los Agentes de IA vía tool calling.

    Convención de rutas: las tablas del Esquema de Backend usan rutas sin
    prefijo (p. ej. /catalog/search); el TRD usa el prefijo /api/v1/... al
    describir el tool calling de los agentes. Este documento resuelve esa
    inconsistencia declarando el prefijo /api/v1 en `servers` y manteniendo los
    `paths` sin prefijo — ambas formas describen el mismo endpoint físico.
  version: "1.0.0"
  contact:
    name: Equipo de Arquitectura de Soluciones — ToolBox JL
  x-sdd:
    canonical: true
    generated_by: "Claude Code (SDD spec-first) — alineado a DOC-BE-01 v1.0 y DOC-ARQ-01 §6.1"
    related_docs:
      - "01_Documento_de_Arquitectura_ToolBoxJL.docx §6.1"
      - "06_Esquema_Backend_ToolBoxJL.docx §3.2, §4"
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /inventory/models:
    post:
      operationId: createToolModel
      summary: Registrar ficha técnica de un nuevo modelo (RF-1.1)
      description: Da de alta un nuevo modelo de herramienta en el catálogo, con sus tarifas y parámetros de garantía y mora.
      tags: [Catálogo e Inventario]
      x-roles: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToolModelInput" }
      responses:
        "201":
          description: Modelo creado.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolModel" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /inventory/units:
    post:
      operationId: createToolUnit
      summary: Alta de unidad física + generación de QR (RF-1.2, HU-13.2)
      description: >
        Registra una unidad física serializada de un modelo existente y
        genera su código QR único (vectorial, ver `qr_code_url`). `numero_serie`
        es el serial del fabricante — HU-13.2 lo muestra junto al QR en la
        vista previa imprimible como el código legible por humanos (no se
        genera un identificador propio adicional tipo "TBJL-DEM-0089": el
        ejemplo del PRD se satisface con el `numero_serie` que ingresa el
        almacenista).
      tags: [Catálogo e Inventario]
      x-roles: [almacenista, admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [modelo_id, numero_serie, fecha_adquisicion, costo_compra, ubicacion_bodega]
              properties:
                modelo_id: { type: string, format: uuid }
                numero_serie: { type: string }
                fecha_adquisicion: { type: string, format: date }
                costo_compra: { type: integer, description: "COP" }
                ubicacion_bodega: { type: string, description: "Ej. 'Estante A3'" }
      responses:
        "201":
          description: Unidad física creada, con su UUID y QR generados.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolUnit" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

    get:
      operationId: listToolUnits
      summary: Panel administrativo de inventario — tabla filtrable (HU-13.1)
      description: >
        Listado paginado de unidades físicas para el panel
        `/logistica/inventario` de `apps/panel-admin`, expandido con
        `modelo_nombre`/`modelo_categoria` (evita que el frontend haga N+1 a
        `GET /catalog/models/{id}` por fila). `estado` en la respuesta es un
        estado DE VISUALIZACIÓN, distinto de `ToolUnit.estado` persistido:
        una unidad con `ToolUnit.estado = Operativo` que tiene un
        `OrderItem` vigente en una Order `confirmada`/`en_curso` se muestra
        acá como `En Alquiler` — ese cruce con Órdenes es el único lugar
        donde existe el concepto "En Alquiler" (no es un valor del enum
        persistido de `ToolUnit.estado`, ver ese schema).
      tags: [Catálogo e Inventario]
      x-roles: [almacenista, admin]
      parameters:
        - name: q
          in: query
          description: Búsqueda libre por código QR (UUID), Serial o nombre de Modelo.
          schema: { type: string }
        - name: estado
          in: query
          description: Filtra por el estado DE VISUALIZACIÓN (ver descripción del endpoint), no el enum persistido.
          schema:
            type: string
            enum: [Operativo, En Alquiler, En Mantenimiento, Dado de Baja]
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Página de unidades físicas filtradas.
          content:
            application/json:
              schema:
                type: object
                required: [items, total, page, pageSize]
                properties:
                  items:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/ToolUnit"
                        - type: object
                          properties:
                            modelo_nombre: { type: string }
                            modelo_categoria: { type: string }
                            estado_visualizacion:
                              type: string
                              enum: [Operativo, En Alquiler, En Mantenimiento, Dado de Baja]
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /inventory/maintenance:
    get:
      operationId: listMaintenanceUnits
      summary: Pestaña "Mantenimiento & Taller" (HU-13.3)
      description: >
        Unidades actualmente `En Mantenimiento` o `Dado de Baja`, con los
        datos de taller de su entrada más reciente en la hoja de vida
        (`ToolUnitStatusLogEntry` — ver los campos nuevos de mantenimiento
        en ese schema, poblados por `PATCH /inventory/units/{id}/status`).
      tags: [Catálogo e Inventario]
      x-roles: [almacenista, admin]
      responses:
        "200":
          description: Unidades en taller o dadas de baja, con su último evento de mantenimiento.
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: "#/components/schemas/ToolUnit"
                    - type: object
                      properties:
                        modelo_nombre: { type: string }
                        ultimo_evento_mantenimiento: { $ref: "#/components/schemas/ToolUnitStatusLogEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /inventory/occupancy:
    get:
      operationId: getInventoryOccupancy
      summary: Ocupación de bodega por ubicación (panel de Almacén)
      description: >
        Cuenta unidades físicas agrupadas por `ubicacion_bodega` (texto
        libre, ver `ToolUnit`) — las unidades sin ubicación asignada se
        agrupan bajo `"Sin ubicación asignada"`. NO incluye capacidad ni
        porcentaje: el sistema no tiene ningún concepto de capacidad total
        de bodega/estante, solo reporta el ocupado real.
      tags: [Catálogo e Inventario]
      x-roles: [almacenista, admin]
      responses:
        "200":
          description: Conteo de unidades por ubicación de bodega.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OcupacionUbicacion" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /inventory/audit-feed:
    get:
      operationId: getInventoryAuditFeed
      summary: Últimos eventos de la hoja de vida de toda la flota (panel de Almacén)
      description: >
        Últimos `limit` (default 10, máx 50) eventos de
        `tool_unit_status_log` de TODAS las unidades, orden cronológico
        descendente, enriquecidos con `numero_serie`/`modelo_nombre`. GAP
        documentado: solo cubre cambios de estado de unidad — NO incluye
        despachos/recepciones de `Order`/`Shipment` (otro bounded context).
      tags: [Catálogo e Inventario]
      x-roles: [almacenista, admin]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
      responses:
        "200":
          description: Últimos eventos de la hoja de vida de la flota.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditoriaEvento" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /inventory/units/{id}:
    get:
      operationId: getToolUnitById
      summary: Ficha de una unidad física al escanear su QR (RF-1.4)
      description: >
        Devuelve la ficha de una unidad física a partir del escaneo de su
        QR. NO incluye la hoja de vida completa (para eso ver
        `GET /inventory/units/{id}/history`) — corregido en Sprint 14, la
        descripción original de este endpoint prometía una "hoja de vida
        resumida" que nunca llegó a declararse en el schema `ToolUnit`.
      tags: [Catálogo e Inventario]
      x-roles: [almacenista, repartidor, admin]
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: Ficha de la unidad.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolUnit" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /inventory/units/{id}/history:
    get:
      operationId: getToolUnitHistory
      summary: Hoja de vida completa de una unidad (HU-13.1, botón "Historial")
      description: >
        Todas las entradas de `tool_unit_status_log` de esta unidad, orden
        cronológico descendente (más reciente primero) — gap detectado en
        Sprint 14 al implementar el botón "Historial" del panel de
        inventario: no existía ningún endpoint para listar la hoja de vida
        completa, solo el estado actual.
      tags: [Catálogo e Inventario]
      x-roles: [almacenista, admin]
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: Hoja de vida completa de la unidad.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ToolUnitStatusLogEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /inventory/units/{id}/status:
    patch:
      operationId: updateToolUnitStatus
      summary: Cambio de estado de una unidad + registro en hoja de vida (RF-1.3, HU-13.3)
      description: >
        Actualiza el estado de desgaste de una unidad física y crea la
        entrada correspondiente en su hoja de vida (tool_unit_status_log).
        `x-roles` amplía a `admin` sobre el contrato original (RF-1.3) para
        cubrir HU-13.3 ("Asignación de una unidad a mantenimiento" /
        "Reintegrar a Servicio" / "Declarar Baja Definitiva" desde el panel
        admin) — el flujo de repartidor/almacenista original sigue igual.
        Los campos de mantenimiento (`tipo_mantenimiento`,
        `falla_reportada`, etc.) solo aplican cuando `estado_nuevo` es
        `En Mantenimiento`; `motivo_baja` solo aplica cuando `estado_nuevo`
        es `Dado de Baja` — el backend no los exige por tipo (quedarían
        `null` en la hoja de vida si se omiten), la UI del panel admin es
        quien los pide como obligatorios en cada caso.
      tags: [Catálogo e Inventario]
      x-roles: [almacenista, repartidor, admin]
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [estado_nuevo]
              properties:
                estado_nuevo:
                  type: string
                  enum: [Nuevo, Excelente, Operativo, En Mantenimiento, Dado de Baja]
                fotos_urls:
                  type: array
                  items: { type: string, format: uri }
                tipo_mantenimiento:
                  type: string
                  enum: [Preventivo, Correctivo]
                falla_reportada: { type: string }
                tecnico_asignado: { type: string }
                costo_estimado: { type: integer, description: "COP" }
                fecha_prevista_fin: { type: string, format: date }
                motivo_baja: { type: string }
      responses:
        "200":
          description: Estado actualizado; entrada de hoja de vida creada.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ToolUnitStatusLogEntry" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /inventory/check-availability:
    get:
      operationId: checkAvailability
      summary: Disponibilidad de un modelo por rango de fechas
      description: >
        Invocado por el Cliente en el flujo de cotización y por los Agentes 2 y 3
        vía tool calling (TRD §4.2, §4.3).
      tags: [Logística]
      x-roles: [cliente, agente-2, agente-3]
      parameters:
        - name: modelo_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: fecha_inicio
          in: query
          required: true
          schema: { type: string, format: date }
        - name: fecha_fin
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Cantidad de unidades disponibles en el rango solicitado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  modelo_id: { type: string, format: uuid }
                  unidades_disponibles: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /orders:
    get:
      operationId: listMyOrders
      summary: Listar mis pedidos (cliente autenticado) — HU-12.1, Fase 3
      description: >
        Pedidos del cliente autenticado (el `cliente_id` sale del JWT, no es
        un parámetro) para la sección "Mis Pedidos Activos" del Home/Catálogo.
        Paginado en bloques de 5 por defecto.
      tags: ["Cotización, Órdenes y Pagos"]
      x-roles: [cliente]
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendiente_pago, confirmada, en_curso, devuelta, cerrada, cancelada]
          description: Sin este filtro, devuelve todos los estados.
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 5 }
      responses:
        "200":
          description: Página de pedidos del cliente, ordenados del más reciente al más antiguo.
          content:
            application/json:
              schema:
                type: object
                required: [items, total, page, pageSize]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Order" }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createOrder
      summary: Crear orden (alquiler o venta)
      description: Crea una orden de alquiler o venta a partir de una cotización aceptada por el cliente.
      tags: ["Cotización, Órdenes y Pagos"]
      x-roles: [cliente]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderInput" }
      responses:
        "201":
          description: Orden creada en estado pendiente_pago.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /orders/{id}:
    get:
      operationId: getOrderById
      summary: Detalle de una orden
      description: Devuelve el detalle completo de una orden, incluidos sus ítems.
      tags: ["Cotización, Órdenes y Pagos"]
      x-roles: [cliente-propietario, admin, gerente]
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: Detalle completo de la orden.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT emitido por Supabase Auth. Los endpoints invocados por agentes de
        IA (x-roles con prefijo "agente-") usan un JWT de servicio con scope
        restringido a las tools que ese agente tiene permitido llamar (TRD §4).
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Solicitud inválida (validación de DTO fallida).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: JWT ausente o inválido.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Rol autenticado sin permiso para esta operación (guard RBAC).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
    ToolModel:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        marca: { type: string }
        categoria: { type: string }
        potencia_w: { type: number }
        peso_kg: { type: number }
        volumen_m3: { type: number }
        tarifa_dia: { type: integer, description: "COP" }
        tarifa_semana: { type: integer, description: "COP" }
        costo_compra: { type: integer, description: "COP" }
        deposito_pct: { type: number }
        interes_mora_dia: { type: number }
        manual_pdf_url: { type: string, format: uri }
        disponible_para_venta: { type: boolean }
        precio_venta:
          type: integer
          nullable: true
          description: >
            COP. Solo aplica si disponible_para_venta=true; null si el
            modelo no está en venta directa (HU-12.2, Fase 3).
    ToolModelInput:
      allOf:
        - $ref: "#/components/schemas/ToolModel"
      required: [nombre, marca, categoria, tarifa_dia]
...
    ToolUnit:
      type: object
      properties:
        id: { type: string, format: uuid }
        modelo_id: { type: string, format: uuid }
        numero_serie: { type: string }
        estado:
          type: string
          enum: [Nuevo, Excelente, Operativo, En Mantenimiento, Dado de Baja]
        fecha_ingreso: { type: string, format: date }
        qr_code_url: { type: string, format: uri }
        fecha_adquisicion:
          type: string
          format: date
          nullable: true
          description: "Sprint 14 (HU-13.2). Nullable — unidades creadas antes de este campo no lo tienen."
        costo_compra:
          type: integer
          nullable: true
          description: "COP. Sprint 14 (HU-13.2)."
        ubicacion_bodega:
          type: string
          nullable: true
          description: "Sprint 14 (HU-13.2), ej. 'Estante A3'."
    ToolUnitStatusLogEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        unidad_id: { type: string, format: uuid }
        estado_anterior:
          type: string
          enum: [Nuevo, Excelente, Operativo, En Mantenimiento, Dado de Baja]
        estado_nuevo:
          type: string
          enum: [Nuevo, Excelente, Operativo, En Mantenimiento, Dado de Baja]
        fotos_urls:
          type: array
          items: { type: string, format: uri }
        autor_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        tipo_mantenimiento:
          type: string
          nullable: true
          enum: [Preventivo, Correctivo]
          description: "Sprint 14 (HU-13.3). Solo poblado cuando estado_nuevo = En Mantenimiento."
        falla_reportada: { type: string, nullable: true, description: "Sprint 14 (HU-13.3)." }
        tecnico_asignado: { type: string, nullable: true, description: "Sprint 14 (HU-13.3)." }
        costo_estimado: { type: integer, nullable: true, description: "COP. Sprint 14 (HU-13.3)." }
        fecha_prevista_fin: { type: string, format: date, nullable: true, description: "Sprint 14 (HU-13.3)." }
        motivo_baja:
          type: string
          nullable: true
          description: "Sprint 14 (HU-13.3). Solo poblado cuando estado_nuevo = Dado de Baja."
    OcupacionUbicacion:
      type: object
      properties:
        ubicacion: { type: string, description: "'Sin ubicación asignada' si la unidad no tiene ubicacion_bodega." }
        cantidad: { type: integer }
    AuditoriaEvento:
      type: object
      description: "ToolUnitStatusLogEntry enriquecido con numero_serie/modelo_nombre, ver GET /inventory/audit-feed."
      properties:
        id: { type: string, format: uuid }
        unidad_id: { type: string, format: uuid }
        numero_serie: { type: string }
        modelo_nombre: { type: string }
        estado_anterior:
          type: string
          nullable: true
          enum: [Nuevo, Excelente, Operativo, En Mantenimiento, Dado de Baja]
        estado_nuevo:
          type: string
          enum: [Nuevo, Excelente, Operativo, En Mantenimiento, Dado de Baja]
        created_at: { type: string, format: date-time }
        autor_id: { type: string, format: uuid }
        falla_reportada: { type: string, nullable: true }
        motivo_baja: { type: string, nullable: true }
...
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        numero_orden:
          type: string
          description: >
            Identificador legible para el cliente (ej. "TJL0000001") — un
            entero autoincremental formateado con prefijo "TJL" y 7 dígitos
            con ceros a la izquierda. `id` (UUID) sigue siendo la clave
            primaria real de la tabla; este campo es solo para mostrarle al
            cliente algo reconocible en vez del UUID (pantalla de "Orden
            Creada Exitosamente", detalle de "Mis Pedidos").
        cliente_id: { type: string, format: uuid }
        tipo: { type: string, enum: [alquiler, venta] }
        estado:
          type: string
          enum: [pendiente_pago, confirmada, en_curso, devuelta, cerrada, cancelada]
        fecha_inicio: { type: string, format: date, nullable: true }
        fecha_fin: { type: string, format: date, nullable: true }
        return_mode:
          type: string
          enum: [en_sede, recogida_domicilio]
        direccion_entrega: { type: string }
        zona_id: { type: string, format: uuid }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
    OrderInput:
      type: object
      required: [tipo, modelo_id]
      properties:
        modelo_id: { type: string, format: uuid }
        tipo: { type: string, enum: [alquiler, venta] }
        fecha_inicio: { type: string, format: date }
        fecha_fin: { type: string, format: date }
        return_mode:
          type: string
          enum: [en_sede, recogida_domicilio]
        direccion_entrega: { type: string }
        zona_id: { type: string, format: uuid }
    OrderItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        order_id: { type: string, format: uuid }
        unidad_id: { type: string, format: uuid }
        tarifa_aplicada: { type: integer, description: "COP" }
        herramienta_nombre:
          type: string
          description: >
            Nombre del modelo de herramienta (ToolModel.nombre) al que
            pertenece unidad_id. Resuelto server-side vía JOIN al leer la
            orden — no se persiste en order_items, así que siempre refleja
            el modelo real de la unidad. Sprint 12 (HU-12.1): el detalle de
            "Mis Pedidos" del cliente lo muestra en vez del identificador
            interno de la unidad (unidad_id no se expone en esa UI, aunque
            sigue en el payload).
...